Move `submit_ogdc_recipe` to new `api` module
Might be worth renaming this eventually. I'm thinking this represents the main
module that defines the priamry high-level Python user API.

THis is more clear because the `submit_ogdc_recipe` function is used for both
current recipe types (and will be used for others in the future).
