+++ src/ogdc_runner/__main__.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import click
 
+from ogdc_runner.api import submit_ogdc_recipe
 from ogdc_runner.argo import get_workflow_status
-from ogdc_runner.recipe.shell import submit_ogdc_recipe
 
 
 @click.group
