+++ src/ogdc_runner/recipe/shell.py
@@ -2,6 +2,8 @@ from __future__ import annotations
 
 import fsspec
 from hera.workflows import (
+    Artifact,
+    Container,
     Steps,
     Workflow,
 )
@@ -14,16 +16,29 @@ from ogdc_runner.argo import (
 
 # Import common utilities
 from ogdc_runner.common import (
-    data_already_published,
-    make_cmd_template,
     make_fetch_input_template,
     make_publish_template,
 )
 from ogdc_runner.constants import SHELL_RECIPE_FILENAME
-from ogdc_runner.exceptions import OgdcDataAlreadyPublished
 from ogdc_runner.models.recipe_config import RecipeConfig
-from ogdc_runner.recipe import get_recipe_config
-from ogdc_runner.recipe.viz_workflow import submit_viz_workflow_recipe
+
+
+def make_cmd_template(
+    name: str,
+    command: str,
+) -> Container:
+    """Creates a command template with an optional custom image."""
+    template = Container(
+        name=name,
+        command=["sh", "-c"],
+        args=[
+            f"mkdir -p /output_dir/ && {command}",
+        ],
+        inputs=[Artifact(name="input-dir", path="/input_dir/")],
+        outputs=[Artifact(name="output-dir", path="/output_dir/")],
+    )
+
+    return template
 
 
 def make_and_submit_shell_workflow(
@@ -86,48 +101,6 @@ def make_and_submit_shell_workflow(
     workflow_name = submit_workflow(w, wait=wait)
 
     return workflow_name
-
-
-def submit_ogdc_recipe(
-    *,
-    recipe_dir: str,
-    wait: bool,
-    overwrite: bool,
-) -> str:
-    """Submit an OGDC recipe for processing via argo workflows.
-
-    Args:
-        recipe_dir: Path to the recipe directory
-        wait: Whether to wait for the workflow to complete
-        overwrite: Whether to overwrite existing published data
-
-    Returns the name of the OGDC shell recipe submitted to Argo.
-    """
-    # Get the recipe configuration
-    recipe_config = get_recipe_config(recipe_dir)
-
-    # Check if the user-submitted workflow has already been published
-    if data_already_published(
-        recipe_config=recipe_config,
-        overwrite=overwrite,
-    ):
-        err_msg = f"Data for recipe {recipe_config.id} have already been published."
-        raise OgdcDataAlreadyPublished(err_msg)
-
-    # Check if the recipe is a visualization workflow
-    if recipe_config.id == "viz-workflow":
-        return submit_viz_workflow_recipe(
-            recipe_dir=recipe_dir,
-            wait=wait,
-        )
-
-    # We currently expect all recipes to be "shell"
-    shell_recipe_workflow_name = make_and_submit_shell_workflow(
-        recipe_config=recipe_config,
-        wait=wait,
-    )
-
-    return shell_recipe_workflow_name
 
 
 def parse_commands_from_recipe_file(recipe_dir: str, filename: str) -> list[str]:
