+++ tests/integration/test_ogdc_recipe.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import pytest
 
+from ogdc_runner.api import submit_ogdc_recipe
 from ogdc_runner.argo import ARGO_WORKFLOW_SERVICE
 from ogdc_runner.exceptions import OgdcDataAlreadyPublished, OgdcInvalidRecipeDir
-from ogdc_runner.recipe.shell import submit_ogdc_recipe
 
 
 def test_submit_ogdc_recipe_with_invalid_dir(tmp_path):
